+++ admiral/cmd/admiral/cmd/root.go
@@ -83,7 +83,7 @@ func GetRootCmd(args []string) *cobra.Command {
 		"The label value, on a namespace, which tells Istio to perform sidecar injection")
 	rootCmd.PersistentFlags().StringVar(&params.HostnameSuffix, "hostname_suffix", "global",
 		"The hostname suffix to customize the cname generated by admiral. Default suffix value will be \"global\"")
-	rootCmd.PersistentFlags().StringVar(&params.WorkloadIdentityLabel, "workload_identity_label", "identity",
+	rootCmd.PersistentFlags().StringVar(&params.LabelSet.WorkloadIdentityLabel, "workload_identity_label", "identity",
 		"The workload identity label key, on deployment which holds identity value used to generate cname by admiral. Default label key will be \"identity\"")
 	loggingOptions.AttachCobraFlags(rootCmd)
